fix(performance) Improve rendering performance in calendar board
Removes need for vue to rerender the tooltips content everytime vue gets
new events. This was problematic when there was more than 1K of
events. It would block the browser.

Turned this process on demand, so for every click on the event a new
tooltip is generated if there none associated to the element. If one already
exists it will use the existing tooltip and show it.

As consequence the code was reduced.